unit tests infile and integration test cleanup
