+++ .gitignore
@@ -25,3 +25,6 @@ target/
 .history
 .ionide
 
+
+# Ignore dev scripts
+scripts/
