+++ src/cli/utils/terminal.rs
@@ -90,3 +90,14 @@ impl Terminal {
         Ok(confirm)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::cli::utils::Terminal;
+    use console::Emoji;
+    #[tokio::test]
+    pub async fn async_command() {
+        let res = Terminal::async_command("white", "file target/release/cardano >/dev/null", Emoji("", "")).await;
+        assert!(res.is_ok());
+    }
+}
