--- tests/cli/mod.rs
@@ -1,2 +1,0 @@
-pub mod cli;
-pub use cli::*;
