--- tests/utils/mod.rs
@@ -1,2 +1,0 @@
-pub mod terminal;
-pub use terminal::*;
